==> notebook/01_iclr_data.ipynb <==
# ICLR Data Processing


v0.1 Jiayao Zhang
June 14, 2023

from __future__ import print_function, absolute_import, division
%load_ext autoreload
%autoreload 2
from pathlib import Path
import numpy as np
import pandas as pd
import tqdm
import json
import uuid
import sqlite3
import seaborn as sns
import matplotlib
from matplotlib import pyplot as plt

sns.set()
sns.set(font_scale=2.5,)
sns.set_style("white")
# sns.set_style({'font.family': 'Libertine'})
sns.set_palette("colorblind")


plt.rcParams.update({'font.size': 20, 'lines.linewidth':4, 'lines.markersize':12, 
                         'xtick.labelsize':15, 'ytick.labelsize':15, 'axes.labelsize':20,
                         'axes.titlesize': 20, 'legend.fontsize':20,
                         'pdf.fonttype': 42, 'ps.fonttype':42, 
                        'image.interpolation':'nearest', 'figure.figsize': (10,8),
                     "text.usetex": False,
         'text.latex.preamble': r"""
        \usepackage{libertine}
        \usepackage[libertine]{newtxmath}
        """,
                     
                    })
tqdm.tqdm.pandas()


## Load Data

### ICLR Database

See (ICLR Datbse)[https://cogcomp.github.io/iclr_database/] for details of getting the data.

import src.data_utils as data_utils

# change the following line accordinlgy
DATA_PATH = Path('path/to/iclr/database')
DB_NAME = "cs_conf_release.db"
LOCAL_DATA_PATH = Path('./data')
CITATION_DATA_PATH = LOCAL_DATA_PATH / 's2_citation'

db_con = sqlite3.connect(DATA_PATH/DB_NAME)
dataloader = data_utils.DataLoader(db_con, None)

# load tables
sub_reviews_agg = dataloader.get_table('sub_reviews_agg')
sub_derive = dataloader.get_table('sub_derive')
sub_arxiv=dataloader.get_table('submission_arxiv')
author_grp = dataloader.get_table('author_grp')



dataloader.get_table('venues').to_csv('venues.csv',header=True,index=False)

sub_reviews_agg.groupby('year').full_decision.unique()

# aggregate the "design matrix"
design_mat = (sub_reviews_agg # submission-review features
 .merge(sub_derive, on=['submission_id','year']) # derived features from submissions
 .merge(author_grp, on=['submission_id','year','arxiv_first']) # arXiv features
)

# merge with Specter embeddings
design_mat = (design_mat
    .merge(pd.read_csv(LOCAL_DATA_PATH / 'submission_cluter_20.csv')[['submission_id','cluster']],
           on='submission_id', how='left')
)

# a lightweight dataframe for quering S2 data
baby_iclr=(design_mat[['submission_id','year','full_decision','arxiv_first',]]
 .merge(subs[['submission_id','title','authors','keywords','tldr']],
       on='submission_id',how='left')
 .merge(sub_arxiv[['submission_id','arxiv_id','primary_category',
                  'published_time','updated_time']], on='submission_id', how='left')
)

baby_iclr.to_csv(LOCAL_DATA_PATH/'baby_iclr.csv', header=True, index=False)

### S2 Citation Data (NCO)

# here the example is 1-year, 3-year, and 5-year (dropping NA rows)
# change the path accordingly
import functools
citation_data_path = [
    ('c365', CITATION_DATA_PATH/'c_365.tsv', '\t'),
    ('c730', CITATION_DATA_PATH/'c_730.tsv', '\t')
    ('c1095', CITATION_DATA_PATH/'c_1095.tsv', '\t'),
    ('c1825', CITATION_DATA_PATH/'c_1825.tsv', '\t')

]

cites_within_window = functools.reduce(
    lambda x,y: x.merge(y,on=['submission_id'], how='outer'),
    [(pd.read_csv(p, sep=s)
      .rename({'cites_within_year':l,'cites_within_window':l},axis=1,errors='ignore')
      [['submission_id',l]]
     )
     for l, p, s in citation_data_path],
)


design_mat = design_mat.merge(cites_within_window, on='submission_id', how='outer')

## Prepare Data for Matching

# log transform some columns
design_mat['log_input_len'] = np.log10(design_mat['input_len'])
design_mat['log_ins_rank_min'] = np.log10(design_mat['ins_rank_min'])
design_mat['log_ins_rank_avg'] = np.log10(design_mat['ins_rank_avg'])
design_mat['log_ins_rank_max'] = np.log10(design_mat['ins_rank_max'])
design_mat['log_author_cite_max'] = np.log10(1+design_mat['author_cite_max'])
design_mat['log_author_cite_avg'] = np.log10(1+design_mat['author_cite_avg'])
design_mat['log_author_cite_min'] = np.log10(1+design_mat['author_cite_min'])

Ns = ['c365', 'c1095', 'c1825']
A = 'arxiv_first'
Y = 'binary_decision'
C_cols = ['year',
 'log_input_len',
 'log_n_tweets',
 'n_fig',
 'n_ref',
 'n_sec',
 'sub_fluency',
 'cluster',
 'n_author',
 'fst_reported_f',
 'any_reported_f',
 'cnt_reported_f',
 'demo_no_us',
 'log_ins_rank_min',
 'log_ins_rank_avg',
 'log_ins_rank_max',
 'log_author_cite_min',
 'log_author_cite_avg',
 'log_author_cite_max'
]

C_vars = ['C(year)',
 'log_input_len',
 'log_n_tweets',
 'n_fig',
 'n_ref',
 'n_sec',
 'sub_fluency',
 'C(cluster)',
 'n_author',
 'C(fst_reported_f)',
 'C(any_reported_f)',
 'cnt_reported_f',
 'C(demo_no_us)',
 'log_ins_rank_min',
 'log_ins_rank_avg',
 'log_ins_rank_max',
 'log_author_cite_min',
 'log_author_cite_avg',
 'log_author_cite_max'
]

## remove NAs in covariates
design_mat = design_mat[~design_mat[C_cols].isna().any(axis=1)]

design_mat.to_csv(LOCAL_DATA_PATH/'design_mat.csv',header=True,index=False)

## Matched Data

(run `02_mathcing.ipynb` at this point.)

def load_matched_df(path, df):
    matched_design_mat = pd.read_csv(path).drop('Unnamed: 0',axis=1)
    matched_design_mat=matched_design_mat[~matched_design_mat.matched_set.isna()]
    return pd.concat([
        matched_design_mat.iloc[::2].assign(grp='treated'),
        matched_design_mat.iloc[1::2].assign(grp='control'),]
    )

## with fine-balance
fb_matched_dat = load_matched_df(LOCAL_DATA_PATH/'fb_matched_design_mat_ordered.csv', design_mat)

cluster_df = design_mat.groupby('cluster').apply(lambda s: 
    s.groupby('primary_keyword').size().sort_values(ascending=False).iloc[:5],

                                   ).to_frame('count').reset_index()

for i in range(20):
    print(cluster_df[cluster_df.cluster==i])

### Table One

design_mat=pd.read_csv(LOCAL_DATA_PATH/'design_mat.csv')

import tableone
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)

def constrcut_tableone(matched, df):
    return tableone.TableOne(pd.concat([matched,
                                      df[~df.arxiv_first].assign(grp='unmatched')]
                            )[C_cols+['grp']].reset_index(), #.assign(grpi=lambda s:s['grp']).set_index(['grpi','submission_id']),
                           categorical=['cluster','demo_no_us','fst_reported_f', 'any_reported_f'],
                           groupby=['grp'],overall=False, smd=True, missing=False,
                           order={'cluster':range(20)},
#                              rename={'control':'Matched Comparison Articles', 'treated': 'Early arXiv\'ed Papers',
#                                     'unmatched': 'Non-Early arXiv\'ed Papers'},
                           pval=False
    )

t1_fb_matched = constrcut_tableone(fb_matched_dat,design_mat).tableone

t1_fb_matched.columns=t1_fb_matched.columns.get_level_values(1)


pd.concat([fb_matched_dat,
            design_mat[~design_mat.arxiv_first].assign(grp='unmatched')]
    )[C_cols+['grp']].groupby('grp').size()

t1_matched = constrcut_tableone(matched_dat,design_mat)

t1_fb_matched = constrcut_tableone(fb_matched_dat,design_mat)

t1_fb_matched_print=(t1_fb_matched[[t1_fb_matched.columns[i] for i in [1,2,0,3,5]]]
  .rename({'control':'Matched Comparison Articles', 
          'treated': 'Early arXiv\'ed Papers (Treated Group)',
            'unmatched': 'Non-Early arXiv\'ed Papers (Unmatched)',
           'SMD (treated,unmatched)': 'SMD (Before Matching)',
           'SMD (control,treated)': 'SMD (After Matching)',
          },axis=1
        )
)

def rename_index(s):
    s = list(s)
    s[0] = r"\texttt{" + s[0].split(',')[0].replace('_','\_')+r"}"
    return tuple(s)

t1_fb_matched_print.index=t1_fb_matched_print.index.map(rename_index)

print(t1_fb_matched_print.to_latex(escape=False))
