==> notebook/03_s2_data_analysis.ipynb <==
# Analysis of S2 Citation Data of ICLR Submissions

(C) David Wadden

import pandas as pd
from pathlib import Path
import json
import numpy as np
from collections import Counter
import os
from pathlib import Path


my_dir = Path(os.getcwd())
data_dir = my_dir.parent / "data"

# ICLR data with S2's and publications dates.
df = pd.read_csv(data_dir / "baby_iclr_ids_and_dates.csv")

# Treat workshop-invited papers as rejections.
df["full_decision_with_workshop"] = df["full_decision"].copy()
df["full_decision"] = df["full_decision"].replace("Invite to Workshop Track", "Reject")


# Citations for all submitted ICLR papers.
citations = [json.loads(line) for line in open(data_dir / "baby_iclr_citations.jsonl")]


## Data checks

Look at distribution of decisions. We can probably just throw out "invite to workshop"; seems like an edge case.

df.full_decision_with_workshop.value_counts()


df.full_decision.value_counts()


len(df)


Distribution of conference years.

df.year.value_counts().sort_index()


df.groupby("year").arxiv_first.value_counts()


Find out how many submissions we missed (i.e. couldn't find S2 ID's for).

**NOTE**: Nearly all the submissions we can't get data on are for rejected papers, since many of these were likely never arXiv'd. We might need to account for this somehow.

- For papers that were accepted, we're missing 1.1%
- For papers that were rejected, we're missing 19.8%. These were likely withdrawn and then never released publicly.

missed = (df["s2_id"] == -1).sum()
print(f"{missed} / {len(df)} submissions missed.")

# Look at this for accepted vs. not.
df_reject = df[df["full_decision"].str.contains("Reject")]
df_accept = df[df["full_decision"].str.contains("Accept")]

missed_reject = (df_reject["s2_id"] == -1).sum()
kept_reject = len(df_reject) - missed_reject
pct_found_reject = round((kept_reject / len(df_reject)) * 100, 1)
print(f"{kept_reject} / {len(df_reject)} ({pct_found_reject}%) rejected submissions found.")

missed_accept = (df_accept["s2_id"] == -1).sum()
kept_accept = len(df_accept) - missed_accept
pct_found_accept = round((kept_accept / len(df_accept)) * 100, 1)
print(f"{kept_accept} / {len(df_accept)} ({pct_found_accept}%) accepted submissions found.")


Look at publication date coverage for papers where we got the S2 ID.

- Nearly all papers have a publication year.
- Roughly 10% of papers don't have a citation date
- Missingness isn't dramatically different for accepted vs. rejected papers.
- **Conclusion**: It should be fine to just toss out the papers with no citation date.

df_s2 = df[df["s2_id"] != -1]
num_papers = len(df_s2)
no_year = df_s2["publication_year"].isna().sum()
no_date = df_s2["publication_date"].isna().sum()

msg = f"{no_year} / {num_papers} papers have no year."
print(msg)

pct_missing = round((no_date / num_papers) * 100, 1)
msg = f"{no_date} / {num_papers} ({pct_missing}%) papers have no exact date."
print(msg)

# Look at whether missing dates are biased toward accepted or rejected papers.
df_s2_reject = df_s2[df_s2["full_decision"].str.contains("Reject")]
df_s2_accept = df_s2[df_s2["full_decision"].str.contains("Accept")]

for name, this_df in [("reject", df_s2_reject), ("accept", df_s2_accept)]:
    num_papers = len(this_df)
    no_date = this_df["publication_date"].isna().sum()
    has_date = num_papers - no_date
    pct_found = round((has_date / num_papers) * 100, 1)
    msg = f"{has_date} / {num_papers} ({pct_found}%) `{name}` papers have an exact date."
    print(msg)


Check that publication date missingness doesn't correlate with preprinting

for val, groupdf in df_s2.groupby("arxiv_first"):
    num_na = groupdf["publication_date"].isna().sum()
    frac_na = round((num_na / len(groupdf)) * 100, 1)
    msg = f"When arXiv is {val}, {frac_na}% of papers have no exact date."
    print(msg)


## Checks on cited papers

Get info on the cited papers. Roughly 15% of cited papers are missing publication dates, but this is virtually identical between accepted vs. rejected papers. So, I think it's probably OK to just throw these out as well.

citations = [json.loads(line) for line in open(data_dir / "baby_iclr_citations.jsonl")]


counts = {"Accept": Counter(), "Reject": Counter()}

for entry in citations:
    if entry["status"] != "Success":
        continue
    else:
        query = f"s2_id == {entry['s2_id']}"
        decision = df.query(query).iloc[0]["full_decision"]
        decision = decision.split(" ")[0]
        if decision not in ["Accept", "Reject"]:
            continue
        cites = entry["citations"]
        for cite in cites:
            if cite["citingPaper"]["publicationDate"] is None:
                counts[decision]["no_date"] += 1
            else:
                counts[decision]["has_date"] += 1

for name in ["Accept", "Reject"]:
    counts_loop = counts[name]
    total = counts_loop["has_date"] + counts_loop["no_date"]
    no_date = counts_loop["no_date"]
    pct_missing = round((no_date / total) * 100, 1)
    msg = f"{no_date} / {total} ({pct_missing}%) `{name}` cited papers have no exact date."
    print(msg)


df.query(f"title == '{entry['title']}' & s2_id == {entry['s2_id']}").iloc[0]["full_decision"]


## Citation analysis

result_dir = Path("../results")
cite_counts = pd.read_table(result_dir / "citations_within_365_include_no_s2orc_true.tsv")


Look at number of citations a year out, comparing papers that were accepted vs. rejected.
Papers that were accepted were cited more.

cite_counts["short_decision"] = [x.split(" ")[0] for x in cite_counts["full_decision"]]
cite_counts.groupby("short_decision")["cites_within_window"].describe()


Look at citations a year out, comparing papers that were arXiv'd vs. not.
It looks like arXiv'ing doesn't make a big difference, at least in mean.

cite_counts.groupby("arxiv_first")["cites_within_window"].describe()


Look at acceptance rate for papers that were arXiv'd vs. not.
Papers that were arXiv'd first are more likely to be accepted.

cite_counts.groupby("arxiv_first").apply(lambda group: (group["short_decision"] == "Accept").mean())
