==> notebook/04_effect_analysis.ipynb <==
# Primary and NOC Analysis

v.1 Jiayao Zhang
June 20

from __future__ import print_function, absolute_import, division
%load_ext autoreload
%autoreload 2
from pathlib import Path
import numpy as np
import pandas as pd
import tqdm
import json
import uuid
import sqlite3

import seaborn as sns
import matplotlib
from matplotlib import pyplot as plt

sns.set()
sns.set(font_scale=2.5,)
sns.set_style("white")
sns.set_palette("colorblind")
tqdm.tqdm.pandas()

import logging
logging.getLogger().setLevel(logging.CRITICAL)
import warnings
warnings.filterwarnings("ignore", )


# Load Processed Data

Ns = ['c365', 'c730', 'c1095']
A = 'arxiv_first'
Y = 'binary_decision'
C_cols = ['year',
 'log_input_len',
 'n_fig',
 'n_ref',
 'n_sec',
 'sub_fluency',
 'cluster',
 'n_author',
 'fst_reported_f',
 'any_reported_f',
 'cnt_reported_f',
 'demo_no_us',
 'log_ins_rank_min',
 'log_ins_rank_avg',
 'log_ins_rank_max',
 'log_author_cite_min',
 'log_author_cite_avg',
 'log_author_cite_max'
]

C_vars = ['C(year)',
 'log_input_len',
 'n_fig',
 'n_ref',
 'n_sec',
 'sub_fluency',
 'C(cluster)',
 'n_author',
 'C(fst_reported_f)',
 'C(any_reported_f)',
 'cnt_reported_f',
 'C(demo_no_us)',
 'log_ins_rank_min',
 'log_ins_rank_avg',
 'log_ins_rank_max',
 'log_author_cite_min',
 'log_author_cite_avg',
 'log_author_cite_max'
]

# change the path accordingly
import functools
citation_data_path = [
    ('c365', './data/s2_citation/c_365.tsv', '\t'),
    ('c730', './data/s2_citation/c_730.tsv', '\t'),
    ('c1095', './data/s2_citation/c_1095.tsv', '\t'),
    ('c1825', './data/s2_citation/c_1825.tsv', '\t')

]

cites_within_window = functools.reduce(
    lambda x,y: x.merge(y,on=['submission_id'], how='outer'),
    [(pd.read_csv(p, sep=s)
      .rename({'cites_within_year':l,'cites_within_window':l},axis=1,errors='ignore')
      [['submission_id',l]]
     )
     for l, p, s in citation_data_path],
)


Load processed "design matrix"

design_mat=pd.read_csv('./data/design_mat.csv')

# stratification by subgroups
design_mat = (
    design_mat.drop(['c365','c1095','c1825'],axis=1)
    .merge(cites_within_window, on='submission_id', how='left')
)

design_mat['inst_level'] =\
    design_mat['log_ins_rank_min'].apply(lambda s : 'L' if s < 1 else ('M' if s < 2 else 'H')
)

design_mat['author_level'] =\
    design_mat['author_cite_avg'].apply(lambda s : 'L' if s < 500 else ('M' if s < 2000 else 'H')
)



def load_matched_df(path, df):
    matched_design_mat = pd.read_csv(path).drop('Unnamed: 0',axis=1)
    matched_design_mat=matched_design_mat[~matched_design_mat.matched_set.isna()]
    return pd.concat([
        matched_design_mat.iloc[::2].assign(grp='treated'),
        matched_design_mat.iloc[1::2].assign(grp='control'),]
    )

Load matched pairs

## with fine-balance on topic cluster
## this will be used in the subsequent analysis
fb_matched_dat = load_matched_df('./data/fb_matched_design_mat_ordered.csv', design_mat)

### Inspecting Sample Sizes

smpsize_df=pd.concat([dat[data_selector(dat,N)]
 .query(f"{N}>={t}")
 .groupby(
    ['arxiv_first','inst_level']).size()
 .to_frame('smps').assign(N=N,t=t)
 for N,t  in thres
])

for i, (N,t) in enumerate(thres):
    q = [50, 75, 90]
    smpsize_df.loc[(smpsize_df.N==N)&(smpsize_df.t==t),'t'] =\
        rf"${q[i%3]}\%$ (${t}$)"

print(
(smpsize_df.reset_index()
 .replace({'inst_level': {
     'H': r'Top-$10$',
     'L': r'Top-$10$ to $100$',
     'M': r'Others'
     },    
    'N':{
         N:r"$\cc^{(" +f'{i+1}' + r")}$"
         for i,N in enumerate(Ns)
        },
    'arxiv_first': {
     True: r"$A=1$",
     False: r"$A=0$",
         }
    })
 .rename({'inst_level':'Author Institution', 'arxiv_first': r'Early arXiving ($A$)'},axis=1)
 .pivot(index=['N','t'], columns=['Author Institution', r'Early arXiving ($A$)'], values='smps')
 .sort_index(axis='columns', level='Author Institution',
            key=lambda s:s.map({r'Top-$10$':0,
                                r'Top-$10$ to $100$':1,
                                r'Others':2})
            )
).to_latex(escape=False)
)


print(
    dat.sort_values(['binary_decision'])
    .replace({'arxiv_first':{
        True: r"$A=1$",
        False: r"$A=0$",
    }})
    .assign(
    acc=lambda s:s['binary_decision'],
    rej=lambda s:s['binary_decision'],
).groupby(['year','arxiv_first']).agg({
    'acc': lambda s : sum(s==1),
    'rej': lambda s : sum(s==0),
        'c365': lambda s: np.round(np.mean(s),2),
        'c730': lambda s: np.round(np.mean(s),2),
    'c1095': lambda s: np.round(np.mean(s),2),
}).to_latex(escape=False)
)

# Primary Analysis

import statsmodels.api as sm
import statsmodels.formula.api as smf

C_formula = '+'.join(C_vars)
formula = f"{A} + " + C_formula
inst_formula = f"{A} + " + '+'.join([var for var in C_vars if 'inst_rank' not in var])
author_formula = f"{A} + " + '+'.join([var for var in C_vars if 'author_cite' not in var])



def print_cond_means_y(predictor, dat, A_col):
    Y_1 = np.mean(predictor.predict(dat[dat[A_col]]))
    Y_0 = np.mean(predictor.predict(dat[~dat[A_col]]))
    return pd.DataFrame([[Y_1,Y_0,Y_1-Y_0]],columns=['Y_1','Y_0','D1'])

### (Optional) Unmatched

Y_unadjust = smf.logit(formula=f"{Y} ~ {formula}", data=design_mat).fit()

print_cond_means_y(Y_unadjust, design_mat, 'arxiv_first')

### Primary Analysis: on the Matched Sample (without NOC)

#### 1. On the full matched sample

Y_matched = smf.logit(formula=f"{Y} ~ {formula}", data=dat).fit()

# helper for bootstraping
def bootstrap_ci(dat, A_col, match_col='matched_set', n_rep=2000, est_name='D1'):
    idx_set = list(dat[match_col].unique())
    
    _bootstraps = []
    for _ in range(n_rep):
        smp_idx = set(np.random.choice(idx_set, len(idx_set), replace=True))
        dt = dat.loc[dat[dat[match_col].isin(smp_idx)].index].reset_index(drop=True)
        
        predictor = smf.logit(formula=f"{Y} ~ {formula}", data=dt).fit(disp=0)
        _bootstraps.append(
            np.mean(predictor.predict(dt[dt[A_col]])) - np.mean(predictor.predict(dt[~dt[A_col]]))
        )
    res = pd.DataFrame(_bootstraps, columns=[est_name])
    res['bidx'] = list(range(n_rep))
    return res

Y_matched_boot=bootstrap_ci(dat, 'arxiv_first',
                            match_col='matched_set', n_rep=2000,)

get_ci_table(Y_matched_boot.assign(grp=0),effect_col='D1', groups=['grp'],alpha=0.05)

#### 2. On subsets of matched sample determined by availability of citation counts

Y_matched_boot_byN=pd.concat([bootstrap_ci(
        dat[data_selector(dat,N)], 'arxiv_first',
        match_col='matched_set', n_rep=2000,est_name='DiD'
    ).assign(N=N)
 for N in tqdm.tqdm(Ns)])

get_ci_table(Y_matched_boot_byN.assign(grp=0),effect_col='DiD', groups=['grp','N'],alpha=0.05)

# Anlaysis with NOC

NOCs are defined by thresholding citation counts using the 0.5, 0.7, and 0.9 -th quantile
in the matched sample. Below is the threshold for each citation count.


_=[print([np.quantile(dat[N].dropna(),q) for q in [0.5,0.75,0.9]])
  for N in ['c365','c730','c1095']
]

thres = [
    ('c365', 4),
    ('c365', 11),
    ('c365', 23),
    ('c730', 10),
    ('c730', 29),
    ('c730', 65),
    ('c1095', 11),
    ('c1095', 41),
    ('c1095', 103),
]
def data_selector(df, N):
    """Select only valid subset of data."""
    if N not in ['c365','c730','c1095']:
        return (df.index >=0) | (df.index<0)
    return df.year <= {
        'c365': 2022,
        'c730': 2021,
        'c1095': 2020,
    }[N]

### Analysis on all matched samples

def fit_full_model(data, N, t):
    """Model for DiD (additive equi-confounding adjustment for DiD)"""
    sel = data_selector(data, N)
    var = f"N_full_{N}_{t}"
    data[var] = 1. * ( data[N] >= t)
    try:
        Y_model = smf.logit(f"{Y}~{formula}", data=data[sel]).fit(disp=0)
    except: 
        Y_model = smf.logit(f"{Y}~{formula}", data=data[sel]).fit(disp=0,method='bfgs')
        
    try:
        N_model = smf.logit(f"{var}~{formula}", data=data[sel]).fit(disp=0)
    except:
        N_model = smf.logit(f"{var}~{formula}", data=data[sel]).fit(disp=0,method='bfgs')
    return Y_model, N_model

def print_cond_means(predictors, dat, A_col):
    """Helper for collecting the results."""
    Y_pred, N_pred = predictors
    X_tr, X_ct = dat[dat[A_col]], dat[~dat[A_col]]
    Y_1 = np.mean(Y_pred.predict(X_tr))
    Y_0 = np.mean(Y_pred.predict(X_ct))
    N_1 = np.mean(N_pred.predict(X_tr))
    N_0 = np.mean(N_pred.predict(X_ct))
    return pd.DataFrame([[
        Y_1, Y_0, N_1, N_0, N_1-N_0, Y_1-Y_0, (Y_1-N_1)-(Y_0-N_0)
    ]], columns=['Y_1','Y_0','N_1','N_0','DN','DY','DiD'])
    print(f"DN: {N_1-N_0:6.4f} vanilla ATET: {Y_1-Y_0:6.4f}\tNOC ATET: {(Y_1-N_1)-(Y_0-N_0):6.4f}")


def cond_means_and_ci(dat, A_col, N, t, match_col='matched_set', n_rep=2000,):
    """Bootstrap CI."""
    idx_set = list(dat[match_col].unique())
    

    _bootstraps = []

    for bidx in range(n_rep):
        smp_idx = set(np.random.choice(idx_set, len(idx_set), replace=True))
        dt = dat[dat[match_col].isin(smp_idx)].reset_index(drop=True)
        sel = data_selector(dt, N)
        predictors = fit_full_model(dt, N, t)
                
        _bootstraps.append(print_cond_means(predictors, dt[sel], A_col)
         .assign(N=N, t=t, n_N1=sum(dt[sel][N] > t))
         .assign(boot_idx=bidx)
        )
    return pd.concat(_bootstraps)


Sample sizes for each study

[dat[data_selector(dat, N)].shape[0] for N in Ns]

#### NOC-adjusted Models

fitted_full_models = [
    fit_full_model(dat, N, t)
    for N, t in tqdm.tqdm(thres)
]

full_model_atet = pd.concat(
    [(print_cond_means(predictors, dat[data_selector(dat, N)], 'arxiv_first')
     .assign(N=N, t=t, n_N1=sum(dat[data_selector(dat, N)][N] > t))
    ) for predictors, (N,t) in tqdm.tqdm(zip(fitted_full_models, thres))
])



full_model_ci = pd.concat([
    cond_means_and_ci(dat, 'arxiv_first', N=N, t=t, n_rep=2000)
    for N, t in tqdm.tqdm(thres)
])

def get_ci_table(boot_df, effect_col='DiD', groups=['N','t'], alpha=0.05):
    ll = alpha / 2.
    hl = 1. - ll
    return pd.concat([
      boot_df.groupby(groups)[effect_col].apply(lambda s : np.mean(s)).to_frame('r'),
      boot_df.groupby(groups)[effect_col].apply(lambda s : np.quantile(s,ll)).to_frame('ll'),
     boot_df.groupby(groups)[effect_col].apply(lambda s : np.quantile(s,hl)).to_frame('hl'),
    ],axis=1)
    

full_model_ci.to_csv('./processed_data/nongrp_bootstrap_smps.csv',
                    header=True, index=False)

full_model_ci_table = get_ci_table(full_model_ci,alpha=0.05)

full_model_ci_table

print(
    r'$'+r'$ & $'.join(
     [f"{r:.4f}"for r in
        [full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").r.iloc[0] 
         for N,t in thres]
     ]
    )+r'$ \\'
)

print(
    r'$'+r'$ & $'.join(
     [f"{r:.4f}"for r in
        [full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").r.iloc[0] 
         for N,t in thres]
     ]
    )+r'$ \\'
)

print(
    r'$'+r'$ & $'.join(
     [f"({ll:.4f},{lh:.4f})"for ll,lh in
        [(full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").ll.iloc[0],
         full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").hl.iloc[0] )
         for N,t in thres]
     ]
    )+r'$ \\'
)

print(
    r'$'+r'$ & $'.join(
     [f"({ll:.4f},{lh:.4f})"for ll,lh in
        [(full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").ll.iloc[0],
         full_model_ci_table.reset_index().query(f"(N=='{N}') and (t=={t})").hl.iloc[0] )
         for N,t in thres]
     ]
    )+r'$ \\'
)

## Stratified Anlaysis for Author Subgroups

def cond_means_ci_subgrp(dat, N, t, A_col, level_col, level_vars, level_lb,
                         all_lb='all',grp_name='level',
                         n_rep=2000, match_col='matched_set'):
    """DiD/NOC for stratified analysis."""

    estimates = []
    for lv, lb in zip([None,]+level_vars, [all_lb,]+level_lb):
        df = dat[dat[level_col] == lv] if lv is not None else dat
        df = df[data_selector(df, N)].reset_index(drop=True)
        idx_set = list(df[match_col].unique())
        
        _bootstraps = []
        for bidx in range(n_rep):
            smp_idx = set(np.random.choice(idx_set, len(idx_set), replace=True))
            dt = df[df[match_col].isin(smp_idx)].reset_index(drop=True)
            try:
                Y_pred, N_pred = fit_full_model(dt, N, t)
            except:
                continue
            X_tr, X_ct = dt[dt[A_col]], dt[~dt[A_col]]
            Y_1 = Y_pred.predict(X_tr)
            Y_0 = Y_pred.predict(X_ct)
            N_1 = N_pred.predict(X_tr)
            N_0 = N_pred.predict(X_ct)
            
            # DiD
            _bootstraps.append(
                (np.mean(Y_1) - np.mean(N_1)) -\
                (np.mean(Y_0) - np.mean(N_0))
            )
        res = pd.DataFrame(_bootstraps, columns=['DiD'])
#         res['bidx'] = list(range(n_rep))
        res['grp'] = grp_name
        res['n'] = len(idx_set)
        res[level_col] = lb
        estimates.append(res)
    return pd.concat(estimates).assign(N=N, t=t)

        
    

def get_ci_table(boot_df, effect_col='DiD', groups=['N','t'], alpha=0.05):
    ll = alpha / 2.
    hl = 1. - ll
    return pd.concat([
      boot_df.groupby(groups)[effect_col].apply(lambda s : np.mean(s)).to_frame('r'),
      boot_df.groupby(groups)[effect_col].apply(lambda s : np.quantile(s,ll)).to_frame('ll'),
     boot_df.groupby(groups)[effect_col].apply(lambda s : np.quantile(s,hl)).to_frame('hl'),
    ],axis=1)
    

author_model_ci = pd.concat([
        cond_means_ci_subgrp(dat, N=N, t=t,A_col='arxiv_first',
            level_col='author_level',level_vars=['L','M','H'],
            level_lb=['<500', '500-2000', '>2000'], grp_name='Max Author Citation',
            n_rep=1000
        )
    for N, t in tqdm.tqdm(thres+[(n,0) for n in Ns])
])

inst_model_ci = pd.concat([
        cond_means_ci_subgrp(dat, N=N, t=t,A_col='arxiv_first',
            level_col='inst_level',level_vars=['L','M','H'],
            level_lb=['top-10', '10-100', 'others'], grp_name='Min Institution Rank',
            n_rep=1000
        )
    for N, t in tqdm.tqdm(thres)
])

full_subgrp_boots = pd.concat([
    inst_model_ci.rename({'inst_level':'level'},axis=1).replace({'level':{'all':'All Institutions'}},),
    author_model_ci.rename({'author_level':'level'},axis=1).replace({'level':{'all':'All Authors'}},),
])


full_subgrp_boots['t_lb']=full_subgrp_boots.apply(lambda s :f"{s['N']}>{s['t']}",axis=1)
full_subgrp_boots['q']=full_subgrp_boots.apply(lambda s :f"{s['N']}>{s['t']}",axis=1)

for i, (N, t) in enumerate(thres):
    q = [r'50\%', r'75\%', r'90\%'][i%3]
    full_subgrp_boots.loc[
        (full_subgrp_boots.N==N)&(full_subgrp_boots.t==t), 'q'] = q
for N in Ns:
        full_subgrp_boots.loc[
        (full_subgrp_boots.N==N)&(full_subgrp_boots.t==0), 'q'] = r'0\%'
        




full_subgrp_boots.to_csv('./stratified_bootstrap_smps.csv',
                         header=True, 
                         index=False)
